sub-modularization
