+++ mongodb_consistent_backup/Backup/Mongodump/__init__.py
@@ -1,0 +1,2 @@
+from Mongodump import Mongodump
+
