+++ mongodb_consistent_backup/Upload/S3/S3Session.py
@@ -4,7 +4,7 @@ from boto import config
 from boto.s3.connection import S3Connection
 
 
-class S3:
+class S3Session:
     def __init__(self, access_key, secret_key, s3_host='s3.amazonaws.com', secure=True, num_retries=5, socket_timeout=15):
         self.access_key     = access_key
         self.secret_key     = secret_key
